backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
google-generativeai
python-dotenv
